Second commit

+++ src/main/java/cat/itacademy/barcelonactiva/bristot/geremias/s05/t01/n01/controllers/BranchController.java
@@ -1,11 +1,11 @@
 package cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.controllers;
 
+import cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.exceptions.BranchNotFoundException;
 import cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.model.domain.Branch;
 import cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.model.dto.BranchDTO;
-import cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.model.services.BranchService;
+import cat.itacademy.barcelonactiva.bristot.geremias.s05.t01.n01.model.services.BranchServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +17,7 @@ import java.util.List;
 public class BranchController {
 
     @Autowired
-    private BranchService branchService;
+    private BranchServiceImpl branchService;
 
     @GetMapping({"/", ""})
     public String getAllBranches(Model model) {
@@ -29,14 +29,18 @@ public class BranchController {
     @GetMapping("/{id}")
     public String getBranch(@PathVariable Integer id, Model model) {
         Branch branch = branchService.getBranch(id);
-        BranchDTO branchDTO = branchService.updateBranch((branch));
+        if (branch == null) {
+            throw new BranchNotFoundException("Branch with ID " + id + " not found");
+        }
+        BranchDTO branchDTO = branchService.updateBranch(branch);
         if (branchDTO != null) {
             model.addAttribute("branch", branchDTO);
             return "branchDetails";
-        }else{
+        } else {
             return "error";
         }
     }
+
 
     @GetMapping("/add")
     public String showAddBranchForm(Model model) {
@@ -78,6 +82,10 @@ public class BranchController {
 
     @PostMapping("/delete/{id}")
     public String deleteBranch(@PathVariable Integer id, Model model) {
+        Branch branch = branchService.getBranch(id);
+        if (branch == null) {
+            throw new BranchNotFoundException("Branch with ID " + id + " not found");
+        }
         branchService.deleteBranch(id);
         return "redirect:/branches";
     }
